feat: support account export commands and support version cmd
