+++ cmd/client_gnfd.go
@@ -13,6 +13,24 @@ import (
 
 const iso8601DateFormatSecond = "2006-01-02T15:04:05Z"
 
+func cmdShowVersion() *cli.Command {
+	return &cli.Command{
+		Name:      "version",
+		Action:    showVersion,
+		Usage:     "print version info",
+		ArgsUsage: "",
+		Description: `
+
+Examples:
+$ gnfd-cmd version  `,
+	}
+}
+
+func showVersion(ctx *cli.Context) error {
+	fmt.Println("Greenfield Cmd Version:", Version)
+	return nil
+}
+
 // NewClient returns a new greenfield client
 func NewClient(ctx *cli.Context) (client.Client, error) {
 	privateKey, _, err := parseKeystore(ctx)
